implemented observer on the product cost to update the product average cost

+++ app/database/migrations/2014_05_21_031612_create_products_table.php
@@ -22,10 +22,10 @@ class CreateProductsTable extends Migration
             $table->mediumText('photos');
             $table->mediumText('description');
 
-            $table->float('price');
-            $table->float('average_cost');
+            $table->decimal('price');
+            $table->decimal('average_cost');
             $table->unsignedInteger('quantity');
-            $table->float('tax', 4, 2);
+            $table->decimal('tax');
 
             $table->timestamps();
