+++ app/database/migrations/2014_05_21_041746_create_product_costs_table.php
@@ -17,7 +17,7 @@ class CreateProductCostsTable extends Migration {
 			$table->increments('id');
             $table->unsignedInteger('product_id');
 
-            $table->float('cost');
+            $table->decimal('cost');
             $table->unsignedInteger('quantity');
 
             $table->timestamps();
